[ORCL-640] Oracle returns invalid data for number type
charBufferPtr acts as a cursor to the charBuffer. When we trim the leading zeros we do not update the charBuffer rather we only move the cursor, charBufferPtr. In other words the charBufferPtr is used to indicated the updated character buffer. And hence when we have to trim the trailing zeros we should use the updated buffer rather than initial buffer

[git-p4: depot-paths = "//Drivers/OracleODBC/Maintenance/1.4/Product/": change = 600770]