[ORCL-641] Add custom type for long binary
Long Binary is a custom type and hence requires a custom type to be implemented within the plugin.

However the only difference between blob and long binary is the Type Name and the userdatatype. Since Type Name is retrieve by the server and userdatatype is only used in the driver level (i.e. simba client will not be utilizing this datatype) we can go ahead and use the blob custom type for long binary as well.

[git-p4: depot-paths = "//Drivers/OracleODBC/Maintenance/1.4/Product/": change = 600911]