[ORCL-645] No data being returned
Issue: The bound data is of type widechar and which need to be converter to char before binding. Since the conversion fails the bound data is invalid and hence no data is being returned.

Solution: Implementing custom typedefault for wide character resolves this issue as the coverter is dependant on typedefault.

[git-p4: depot-paths = "//Drivers/OracleODBC/Maintenance/1.4/Product/": change = 611864]